Refactor init-task API
