+++ fractal_server/app/runner/v2/task_output.py
@@ -1,7 +1,6 @@
 from typing import Optional
 
 from pydantic import BaseModel
-from pydantic import Field
 
 from ....images import SingleImage
 from .models import DictStrAny
@@ -13,21 +12,6 @@ class TaskOutput(BaseModel):
     removed_images: Optional[list[SingleImage]] = None
     new_attribute_filters: Optional[DictStrAny] = None
     new_flag_filters: Optional[dict[str, bool]] = None
-
-    class Config:
-        extra = "forbid"
-
-
-class InitArgsModel(BaseModel):
-    class Config:
-        extra = "forbid"
-
-    path: str
-    init_args: DictStrAny = Field(default_factory=dict)
-
-
-class InitTaskOutput(BaseModel):
-    parallelization_list: list[InitArgsModel] = Field(default_factory=list)
 
     class Config:
         extra = "forbid"
